+++ migrations/005_create_orders.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS orders (
     table_id UUID NOT NULL REFERENCES tables(id) ON DELETE CASCADE,
     session_token VARCHAR(64) NOT NULL UNIQUE,
     status VARCHAR(20) NOT NULL CHECK (status IN ('open', 'checkout_pending', 'paid', 'closed')),
-    total_amount DECIMAL(10,2) NOT NULL DEFAULT 0.00,
+    total_amount BIGINT NOT NULL DEFAULT 0,
     payment_method VARCHAR(20),
     stripe_session_id TEXT,
     opened_by UUID REFERENCES users(id) ON DELETE SET NULL,
